deleted:    src/Server.js 	new file:   src/game/MainScene.js 	deleted:    src/game/config/MainScene.js 	new file:   src/game/entities/bullet.js 	new file:   src/game/entities/player.js 	modified:   src/game/game.js 	deleted:    src/main.js
