+++ src/game/game.js
@@ -1,6 +1,6 @@
 import { useEffect } from "react";
 import Phaser from "phaser";
-import MainScene from "./config/MainScene";
+import MainScene from "./MainScene";
 
 export default function Game(props) {
    const {server} = props;
